code refactoring: create sub components and style component
